Refactored health check code such that module and feature files are separated. Updated references in the tests as well.
