+++ src/Microsoft.Health.Api.UnitTests/Features/HealthCheck/HealthCheckCachingTests.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Logging.Abstractions;
-using Microsoft.Health.Api.Modules.HealthChecks;
+using Microsoft.Health.Api.Features.HealthChecks;
 using Microsoft.Health.Core.Internal;
 using Microsoft.Health.Test.Utilities;
 using NSubstitute;
